fixed showing wrong side so everything got culled

+++ Engine/src/Camera.cpp
@@ -23,7 +23,7 @@ namespace sa {
     {
         m_viewport.offset = { 0, 0 };
         m_viewport.extent = windowExtent;
-        m_apectRatio = m_viewport.extent.width / m_viewport.extent.height;
+        m_apectRatio = (float)m_viewport.extent.width / m_viewport.extent.height;
         m_projMat = glm::perspective(m_fov, m_apectRatio, m_near, m_far);
 
     }
@@ -40,7 +40,7 @@ namespace sa {
 
     void Camera::setViewport(Rect viewport) {
         m_viewport = viewport;
-        m_apectRatio = m_viewport.extent.width / m_viewport.extent.height;
+        m_apectRatio = (float)m_viewport.extent.width / m_viewport.extent.height;
         m_projMat = glm::perspective(m_fov, m_apectRatio, m_near, m_far);
     }
